+++ src/f_cli/cli.py
@@ -1,10 +1,14 @@
 """f-cli entrypoint."""
-from typing import Any, Dict, List, Optional, Union
+from typing import Any, Dict, List, Optional, Tuple, Union
+
 import sys
 
 import click
 
 from . import __version__
+from .logging import LogLevels, setup_logging
+from .resources.click_classes import CliGroup
+from .resources.click_options import click_common_options
 
 
 CLICK_CONTEXT_SETTINGS: Dict[str, Union[int, List[str]]] = {
@@ -31,11 +35,23 @@ def print_version(ctx: Optional[click.Context],
     ctx.exit()
 
 
-@click.group(context_settings=CLICK_CONTEXT_SETTINGS)
+@click.group(context_settings=CLICK_CONTEXT_SETTINGS, cls=CliGroup)
 @click.option('--version', is_flag=True, callback=print_version,
               expose_value=False, is_eager=True)
-def cli() -> None:
+@click_common_options
+@click.pass_context
+def cli(ctx: click.Context,
+        debug: bool = False,
+        verbose: bool = False) -> None:
     """f-cli tool."""
+    args: Tuple[str, ...] = getattr(ctx.obj, 'args', ctx.obj)
+
+    if debug or '--debug' in args:
+        setup_logging(LogLevels.DEBUG)
+    elif verbose or '--verbose' in args:
+        setup_logging(LogLevels.VERBOSE)
+    else:
+        setup_logging(LogLevels.INFO)
 
 
 def main() -> int: